created Components folder & seperated App components into DataList component & RenderItem component
