+++ Components/DataList.js
@@ -1,0 +1,123 @@
+const listData = [
+  {
+    id: 1,
+    name: 'Amy Farha',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/ladylexy/128.jpg',
+    subtitle: 'Vice President',
+  },
+  {
+    id: 2,
+    name: 'Chris Jackson',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/adhamdannaway/128.jpg',
+    subtitle: 'Vice Chairman',
+  },
+  {
+    id: 3,
+    name: 'Farha Amy',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/ladylexy/128.jpg',
+    subtitle: 'Manager',
+  },
+  {
+    id: 4,
+    name: 'Sally John',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/ladylexy/128.jpg',
+    subtitle: 'Team Lead',
+  },
+  {
+    id: 5,
+    name: 'Don Boss',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/ladylexy/128.jpg',
+    subtitle: 'Team supervisor',
+  },
+  {
+    id: 6,
+    name: 'John Honnay',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/ladylexy/128.jpg',
+    subtitle: 'Team manager',
+  },
+  {
+    id: 7,
+    name: 'Amy watson',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/ladylexy/128.jpg',
+    subtitle: 'Team Lead',
+  },
+  {
+    id: 8,
+    name: 'Shakira',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/ladylexy/128.jpg',
+    subtitle: 'Quality analyst',
+  },
+  {
+    id: 9,
+    name: 'Ola',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/ladylexy/128.jpg',
+    subtitle: 'Tester',
+  },
+  {
+    id: 10,
+    name: 'Stephan',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/ladylexy/128.jpg',
+    subtitle: 'senior programmer',
+  },
+  {
+    id: 11,
+    name: 'Jackson',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/adhamdannaway/128.jpg',
+    subtitle: 'senior programmer',
+  },
+  {
+    id: 12,
+    name: 'Allan Jackson',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/adhamdannaway/128.jpg',
+    subtitle: 'mid programmer',
+  },
+  {
+    id: 13,
+    name: 'Jack sta',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/adhamdannaway/128.jpg',
+    subtitle: 'mid programmer',
+  },
+  {
+    id: 14,
+    name: 'Briton',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/adhamdannaway/128.jpg',
+    subtitle: 'Regular programmer',
+  },
+  {
+    id: 15,
+    name: 'Antony',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/adhamdannaway/128.jpg',
+    subtitle: 'Regular programmer',
+  },
+  {
+    id: 16,
+    name: 'Maagda',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/adhamdannaway/128.jpg',
+    subtitle: 'Junior programmer',
+  },
+  {
+    id: 17,
+    name: 'Karolina',
+    avatar_url:
+      'https://s3.amazonaws.com/uifaces/faces/twitter/adhamdannaway/128.jpg',
+    subtitle: 'Junior programmer',
+  },
+];
+
+export default listData;
